Time: 2 ms (33.82%), Space: 6.6 MB (13.31%) - LeetHub

+++ 0342-power-of-four/0342-power-of-four.cpp
@@ -8,7 +8,7 @@ public:
     bool isPowerOfFour(int n) {
 
         if(n == 0) return false;
-        
-        return double_is_int((log10(n) / (double)log10(4)));
+
+        return double_is_int((log10(n) / log10(4)));
     }
 };
